% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighborhoods_utils.R
\name{neighborhoods_to_matrix}
\alias{neighborhoods_to_matrix}
\title{Convert neighborhood column to other format}
\usage{
neighborhoods_to_matrix(
  data,
  fit = NULL,
  cell_names = NULL,
  neighborhood_column_name = "neighborhood",
  id_column_name = "name",
  return_sparse = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{the 'neighborhood' column from the output of `lemur::find_de_neighborhoods`
or the full data.frame}

\item{fit}{the `lemur_fit` object (optional). The column names of fit are used
as the set of possible values in the neighborhood columns.}

\item{cell_names}{a character vector with the set of possible values
in the neighborhood columns.}

\item{neighborhood_column_name, id_column_name}{if `data` is a data.frame,
these identifiers are used to extract the relevant columns}

\item{return_sparse}{if `TRUE` a `dgCMatrix` is returned. Otherwise a regular
dense matrix is returned.}

\item{verbose}{indicator if additional messages are printed.}
}
\value{
a (sparse) matrix with `length(data)` rows (assuming  `data` is the neighborhood
  column) and one column for each element of `cell_names` (or the set of all cell labels
  occurring in `data`).
}
\description{
Convert neighborhood column to other format
}
