% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_de.R
\name{scale_color_de}
\alias{scale_color_de}
\alias{scale_colour_de}
\alias{scale_fill_de}
\alias{scale_de}
\title{A divergent color scale designed to display differential expression values}
\usage{
scale_color_de(...)

scale_colour_de(...)

scale_fill_de(...)

scale_de(
  name = ggplot2::waiver(),
  ...,
  qlimits = c(0, 1),
  low = munsell::mnsl("10B 4/6"),
  mid = munsell::mnsl("N 8/0"),
  high = munsell::mnsl("10R 4/6"),
  midpoint = 0,
  space = "Lab",
  na.value = "grey50",
  oob = scales::squish,
  transform = "identity",
  guide = "colourbar",
  aesthetics = "colour"
)
}
\arguments{
\item{qlimits}{the quantile limits. Alternative to limits, which allows
specification of the quantiles to be displayed. By default, it is
combined with \code{oob = scales::squish} to squish outliers into a
the desired range. The argument has to be either \code{NULL}, in which
case it is ignored, a numeric vector with two elements between 0 and 1,
or a numeric vector with only a single element, in which case
the limits are \code{c(qlimits, 1-qlimits)}.}
}
\description{
Creates a divergent colorscale around the \code{midpoint}. Unlike the
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradient2()}} blue signals low values and red
are high values. In addition, \code{qlimits} filters out outliers that
distort the overall picture.
}
\examples{
  library(ggplot2)
  library(tibble)

  dat <- tibble(x = 1:10, y = seq(-3, 5, length.out = 10))
  ggplot(dat, aes(x = x, y = y)) +
    geom_col(aes(fill = y)) +
    scale_fill_de()

  ggplot(dat, aes(x = x, y = y)) +
    geom_col(aes(fill = y)) +
    scale_fill_de(qlimits = c(0, 0.7))

  dat <- tibble(x = 1:10, y = seq(-1, 3, length.out = 10))
  ggplot(dat, aes(x = x, y = exp(y))) +
    geom_col(aes(fill = y)) +
    scale_fill_de(midpoint = 1, qlimits = c(0.5, 1))

}
\seealso{
This function is based on \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradient2()}}.
See there, for details about the other parameter.
}
