% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_longer.R
\name{sce_pivot_longer}
\alias{sce_pivot_longer}
\alias{fit_pivot_longer}
\title{Convert a \code{lemur_fit} into a tidy datastructure}
\usage{
sce_pivot_longer(
  sce,
  genes = NULL,
  cells = NULL,
  assays = 1,
  reduced_dims = NULL,
  rownames = NULL,
  colnames = NULL
)

fit_pivot_longer(
  fit,
  genes = NULL,
  cells = NULL,
  assays = fit$use_assay,
  reduced_dims = "embedding",
  rownames = NULL,
  colnames = NULL
)
}
\arguments{
\item{sce}{a \code{SingleCellExperiment}}

\item{genes}{a selection of rows (i.e., genes) from \code{fit}. It can either
be an index vector, a character vector with \code{rownames} or a
\code{dplyr::filter}-like expression wrapped in \code{vars()} that can refer
to columns of the \code{rowData(fit)}. Default: \code{NULL} which means no genes
are included.}

\item{cells}{a selection of columns (i.e., cells) from \code{fit}. It can either
be an index vector, a character vector with \code{colnames} or a
\code{dplyr::filter}-like expression wrapped in \code{vars} that can refer to
columns of the \code{colData(fit)}. Default: \code{NULL} which means that all
cells are included.}

\item{assays}{a character vector with the names of assays whose
values are included in the output. To see all options run \code{assayNames(fit)}.}

\item{reduced_dims}{a character with the names of the \code{reducedDim(fit)}
which are included as \code{matrix} columns in the output. To see all options
run \code{reducedDimNames(fit)}.}

\item{rownames, colnames}{string that specifies the column name for
for the \code{rownames(sce)} and \code{colnames(sce)} respectively. Default:
\code{NULL} which means no extra columns are included}

\item{fit}{the \code{lemur_fit} produced by \code{lemur()}}
}
\value{
a \code{tibble} with the content from
\itemize{
\item \code{colnames(fit)} if \code{colnames} is not \code{NULL},
\item \code{colData(fit)} subset according to \code{cells},
\item \code{rownames(fit)} if \code{rownames} is not \code{NULL},
\item \code{rowData(fit)}subset according to \code{genes},
\item one numeric column per entry in \code{assays}.
\item one matrix-column per entry in \code{reducedDims}.
}
}
\description{
The function takes a \code{lemur_fit} (or any other \code{SingleCellExperiment}) and
converts it to a tidy tibble.
}
